Exercise 4.8 solution + refactor
