+++ README.md
@@ -172,3 +172,15 @@ If there are many top bloggers, then it is enough to show any one of them.
 
 **Solution:**
 Implemented the test.
+
+## Exercise 4.8: Blog list tests, step1
+**Task:**
+Use the supertest package for writing a test that makes an HTTP GET request to the /api/blogs url. Verify that the blog list application returns the correct amount of blog posts in the JSON format.
+
+Once the test is finished, refactor the route handler to use the async/await syntax instead of promises.
+
+Notice that you will have to make similar changes to the code that were made in the material, like defining the test environment so that you can write tests that use their own separate database.
+
+**Solution:**
+Implemented in bloglist-backend application and the application refactored.
+
