+++ bloglist-backend/controllers/blogs.js
@@ -1,20 +1,23 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 
-blogsRouter.get('/', (request, response, next) => {
-  Blog
-    .find({})
-    .then(blogs => {
+blogsRouter.get('/', async (request, response, next) => {
+    try {
+      const blogs = await Blog.find({})
       response.json(blogs)
-    }).catch(error => next(error))
+    } catch (error) {
+      next(error)
+    }
 })
 
-blogsRouter.post('/', (request, response, next) => {
-  const blog = new Blog(request.body)
-
-  blog.save().then(result => {
-    response.status(201).json(result)
-  }).catch(error => next(error))
+blogsRouter.post('/', async (request, response, next) => {
+  try {
+    const blog = new Blog(request.body)
+    const newBlog = await blog.save()
+    response.status(201).json(newBlog)
+  } catch (error) {
+    next(error)
+  }
 })
 
 module.exports = blogsRouter
