+++ bloglist-backend/package.json
@@ -4,14 +4,16 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "jest --verbose",
-    "dev": "nodemon index.js"
+    "start": "NODE_ENV=production node index.js",
+    "dev": "NODE_ENV=development nodemon index.js",
+    "test": "NODE_ENV=test jest --verbose --runInBand --forceExit"
   },
   "author": "Dominik Rohal",
   "license": "ISC",
   "devDependencies": {
     "jest": "^29.0.3",
-    "nodemon": "^2.0.20"
+    "nodemon": "^2.0.20",
+    "supertest": "^6.2.4"
   },
   "dependencies": {
     "cors": "^2.8.5",
