--- bloglist-backend/tests/test.test.js
@@ -1,8 +1,0 @@
-const listHelper = require('../utils/list_helper')
-
-test('dummy returns one', () => {
-  const blogs = []
-
-  const result = listHelper.dummy(blogs)
-  expect(result).toBe(1)
-})
